# companion/openapi.yaml
openapi: 3.1.0
info:
  title: Companion API
  version: 0.1.0
  description: |
    API del empleado IA central del ecosistema.

    Companion acepta API key o Bearer JWT interno. Cuando hay principal
    autenticado, `org_id` se interpreta como customer org/work context: la
    empresa cliente donde trabaja el empleado IA. Headers como `X-Org-ID` y
    `X-User-ID` quedan como compatibilidad de dev/test, no como fuente canonica
    productiva.
servers:
  - url: http://localhost:18085
security:
  - apiKeyAuth: []
  - bearerAuth: []
paths:
...
  /v1/employee-profiles:
    get:
      summary: Lista Employee Profiles
      description: |
        Superficie publica recomendada para perfiles de Virtual Employees.
        En v1 reutiliza Agent Profiles internamente; `profile_id` es el UUID
        publico del perfil y `profile_key` conserva la key tecnica usada por
        Agent/Runtime durante la transicion.
      parameters:
        - in: query
          name: lifecycle
          schema: { type: string, enum: [active, archived, trash, all], default: active }
        - in: query
          name: include_archived
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Employee Profile list
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee_profiles:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeeProfile" }
                  profiles:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeeProfile" }
    post:
      summary: Crea un Employee Profile
      description: Crea un perfil con UUID publico automatico; el cliente puede enviar `profile_key` humano o dejarlo autogenerado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeProfileInput" }
      responses:
        "201":
          description: Employee Profile created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeProfile" }
  /v1/employee-profiles/{profile_id}:
    get:
      summary: Obtiene un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Employee Profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeProfile" }
        "404": { description: Employee Profile not found }
    patch:
      summary: Actualiza un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeProfileInput" }
      responses:
        "200":
          description: Employee Profile updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeProfile" }
  /v1/employee-profiles/{profile_id}/status:
    post:
      summary: Cambia el status de un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, archived, trashed] }
      responses:
        "200":
          description: Employee Profile status changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeProfile" }
  /v1/employee-profiles/{profile_id}/archive:
    post:
      summary: Archiva un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Employee Profile archived }
  /v1/employee-profiles/{profile_id}/restore:
    post:
      summary: Restaura un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Employee Profile restored }
  /v1/employee-profiles/{profile_id}/trash:
    post:
      summary: Mueve un Employee Profile a papelera
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Employee Profile trashed }
  /v1/employee-profiles/{profile_id}/purge:
    delete:
      summary: Elimina definitivamente un Employee Profile en papelera
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Employee Profile purged }
  /v1/employee-profiles/{profile_id}/versions:
    get:
      summary: Lista snapshots anteriores de un Employee Profile
      parameters:
        - in: path
          name: profile_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Employee Profile versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeeProfileVersion" }
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    EmployeeProfileInput:
      type: object
      required: [name, system_prompt]
      properties:
        profile_key:
          type: string
          description: Human-readable technical key. Optional on create; generated from name/family/version when omitted.
        family_id: { type: string }
        version_label: { type: string, default: v1 }
        name: { type: string }
        description: { type: string }
        system_prompt: { type: string }
        max_autonomy: { type: string, enum: [A0, A1, A2, A3, A4, A5], default: A1 }
        default_capability_ids:
          type: array
          description: Target model field. During transition it maps to the existing profile capability allowlist.
          items: { type: string, format: uuid }
        allowed_capabilities:
          type: array
          deprecated: true
          description: Legacy capability keys used by the technical runtime profile.
          items: { type: string }
        allowed_tools:
          type: array
          deprecated: true
          description: Legacy direct tool allowlist. Employees should reference capabilities instead.
          items: { type: string }
        memory_policy: { type: object, additionalProperties: true }
        llm_config: { type: object, additionalProperties: true }
        enabled: { type: boolean, default: true }
    EmployeeProfile:
      allOf:
        - $ref: "#/components/schemas/EmployeeProfileInput"
        - type: object
          required: [profile_id, profile_key, name, system_prompt, max_autonomy, status]
          properties:
            id: { type: string, format: uuid }
            profile_id:
              type: string
              format: uuid
              description: Public EmployeeProfile UUID.
            profile_key:
              type: string
              description: Compatibility key currently used by the technical runtime profile.
            status: { type: string, enum: [active, disabled, archived, trashed] }
            archived_at: { type: string, format: date-time }
            trashed_at: { type: string, format: date-time }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    EmployeeProfileVersion:
      allOf:
        - $ref: "#/components/schemas/EmployeeProfile"
        - type: object
          properties:
            employee_profile_id: { type: string, format: uuid }
            original_created_at: { type: string, format: date-time }
            original_updated_at: { type: string, format: date-time }
            saved_at: { type: string, format: date-time }
...